выносит работу с окном в отдельный модуль
